requirements.txt
Django==4.2
djangorestframework==3.14
djangorestframework-simplejwt==5.2.0
django-cors-headers==4.0
django-filter==23.1
python-decouple==3.8
psycopg2-binary==2.9.6
python-dotenv==1.0.0
